+++ vue-todo/src/store/store.js
@@ -1,5 +1,7 @@
 import Vue from "vue";
 import Vuex from "vuex";
+import * as getters from './getters';
+import * as mutations from './mutations';
 
 Vue.use(Vuex);
 
@@ -21,30 +23,6 @@ export const store = new Vuex.Store({
     state: {
         todoItems: storage.fetch()
     },
-    getters: {
-        storedTodoItems(state){
-            return state.todoItems;
-        }
-    },
-    mutations: {
-        addOneItem(state, todoItem) {
-            const obj = { completed: false, item: todoItem };
-            localStorage.setItem(todoItem, JSON.stringify(obj));
-            state.todoItems.push(obj);
-        },
-        removeOneItem(state, payload) {
-            localStorage.removeItem(payload.todoItem.item);
-            state.todoItems.splice(payload.index, 1); // splice : 삭제 후 새로운 배열 반환
-        },
-        toggleOneItem(state, payload) {
-            // todoItem.completed = !todoItem.completed; // 안티 패턴
-            state.todoItems[payload.index].completed = !state.todoItems[payload.index].completed;
-            localStorage.removeItem(payload.todoItem.item);
-            localStorage.setItem(payload.todoItem.item, JSON.stringify(payload.todoItem));
-        },
-        clearAllItems(state) {
-            localStorage.clear();
-            state.todoItems = [];
-        }
-    }
+    getters,
+    mutations
 });
